Add functionality for iding connections and map_variable elements.
The functionality to id connections and the map variables in the connection has been added.  To unset one of these ids you must set the id to the empty string. hasEquivalentVariable now checks to make sure the share_ptr is still valid. Minor tidy up of commented out code and test values.