Validator tests coded
Validator model is coded with xrefitem links that doxygen will use to generate a history of the validation coverage against the draft CellML2.0 specifications.  NB: Aliases need to be added *after* project configuration as DOXYGEN_ALIASES is not working yet.  See comments at start of validation.cpp or inside CMakeLIsts for text to add to Doxyfile.