Generator: make sure that we have 100% coverage.
The origin comment code is always going to be the first bit of code, so no need to check whether we first need to add an empty line.

FWIW, the idea was in case we ever wanted to switch things around, but when it comes to the origin comment, it's probably the only thing that we will always want to have first, hence no need for that test.