Generator: reworked the generation of the interface declaration for `LIBCELLML_VERSION`, `STATE_COUNT`, `VARIABLE_COUNT`, `VOI_INFO`, `STATE_INFO` and `VARIABLE_INFO`.
Indeed, the original plan was not to have both an interface and an implementation string for something, but just have one and then rely on a declaration string to distinguish between the interface and implementation. However, this approach doesn't work with functions or, rather, makes it unnecessarily complex because of the C and Python syntaxes being quite different (and, no doubt, we would have the same issue with other languages), thus making our "generic" code rather cumbersome.