Code generator: always return a GeneratorVariablePtr for our VOI, states and variables.
Also, GeneratorVariable now keeps track of the Component object associated with a Variable object. This is related to issue #546 where the idea is that the Variable object is the one where a variable is initialised while the Component object is the component under which we would expect the variable to be listed (e.g. in OpenCOR). That Component object maybe the same as the Variable object's parent or not. It all depends on whether the reference variable is in the same component as where it is computed (this is only valid for an ODE system since a DAE system may use several equations to compute a variable, with those different equations in different components...!).