Analyser/Generator: use weak pointers where appropriate.
Note that when it comes to the Python bindings for our Generator class, we cannot have a default Python profile since a Generator object now uses a weak pointer to keep track of a profile.

So, what wwe used to do in src/bindings/interface/generator.i will create a profile and assign it to our Generator object, but will then destroy that profile, so our Generator object doesn't have a profile anymore.

So, no matter which binding we use, the default profile will always be the C profile.