Demonstrate a potential bug in the testing for cyclical imports.
Adding a new test which has two models which import the same common units definitions and then one
of the models also imports the other model. This results in an cyclical import error reported when
the importer resolves the imports. The error message is wrong in that it refers to components as
units; and also wrong in that as far as I can tell there is no cycle in the imported units.

The test models added are the minimal demonstration of this that I've reduced from an actual model.
Reducing further (e.g., fewer variables, or only importing a subset of the units) results in different
behaviour, which is possibly a different issue.