Analyser: handle external variables as soon as possible.
We used to do handle them after having analysed the model which meant having to requalify a variable-based constant equation as an algebraic equation if it depended on one or several external variables. In other words, as a developer, we had to think about what consequence(s) an external variable might have on our analysed model. Until now, it wasn't too much a problem, but with our wish to support systems of non-linear algebraic equations (and then DAEs), this is going to become a nightmare.