Added a macro (!!) to test shared pointers.
Although I hate macros, I couldn't find a better way to address the fact that we want to check that a shared pointer exists before using it and, yet, not have llvm-cov "complain" to us. Indeed, we want to be able to do something like:

auto sharedPtr = weakPtr.lock();

if (sharedPtr != nullptr) {
    ...
}

In 99.9% of cases, such a test will always be true, which means that llvm-cov will complain about a lack of branch coverage while it's just us being (rightly) careful by making sure that weakPtr still holds a valid reference to an object before using it.