Tests: updated the algebraic_system_with_various_dependencies test.
'd' was originally classified as a computed constant (since 'b' and 'c' have an initial value, but when we classify 'd', we don't yet know about the final state of 'b' and 'c'). 'b' and 'c' end up being algebraic variables that are computed using an NLA system, hence 'd' should be classified as an algebraic variable in the end, and 'd' have a dependency on 'b' and 'c'.