Generator: initialise variables that need to be computed using an NLA system.
Such variables are the only unknown variable in an equation, but they are not on their own on either the LHS or RHS of that equation, hence we solve them using an NLA system. This means that we need to provide an initial guess and we can safely use zero since that NLA system has only one solution.