Generator: don't initialise our external variables in initialiseVariables().
There was never a need to initialise them there in the first place. In fact, an external variable is effectively an algebraic variable which is computed by an end user. So, it needs to be computed in computeRates() (if available) and in computeVariables().

Also fixed an issue with initialiseVariables() not always including the VOI.