CMake: don't use hints to find llvm-cov and llvm-profdata.
Someone might have a version of llvm-cov and llvm-profdata that they use and which is not located in /Library/Developer/CommandLineTools/usr/bin. This means that they will need to provide LLVM_BIN_DIR which is not ideal, not least since there are PREFERRED_LLVM_COV_NAMES and PREFERRED_LLVM_PROFDATA_NAMES to play with, if needed.