Analyser: account for the case where a variable is underconstrained.
Indeed, we check whether the number of equations in an NLA system is greater than the number of unknowns and, if so, consider the model as overconstrained. So, we should also do the reverse, i.e. check whether the number of equations in an NLA system is lower than the number of unknowns, in which case the model should be considered underconstrained.